+++ monkey/monkey_island/cc/next_ui/src/_components/getting-started/GettingStarted.tsx
@@ -1,0 +1,9 @@
+'use client';
+
+import Box from '@mui/material/Box';
+
+const GettingStarted = () => {
+    return <Box>Hello There!</Box>;
+};
+
+export default GettingStarted;