+++ monkey/monkey_island/cc/next_ui/src/app/(protected)/page.tsx
@@ -1,0 +1,10 @@
+import classes from './page.module.scss';
+import GettingStarted from '@/_components/getting-started/GettingStarted';
+
+export default function AppHome() {
+    return (
+        <div id={classes['app-root-page']}>
+            <GettingStarted />
+        </div>
+    );
+}