+++ monkey/monkey_island/cc/next_ui/src/app/error.tsx
@@ -1,0 +1,28 @@
+'use client';
+
+import { useEffect } from 'react';
+
+export default function Error({
+    error,
+    reset
+}: {
+    error: Error & { digest?: string };
+    reset: () => void;
+}) {
+    useEffect(() => {
+        console.error(error);
+    }, [error]);
+
+    return (
+        <div>
+            <h2>Something went wrong!</h2>
+            <button
+                onClick={
+                    // Attempt to recover by trying to re-render the segment
+                    () => reset()
+                }>
+                Try again
+            </button>
+        </div>
+    );
+}