+++ monkey/monkey_island/cc/next_ui/src/constants/http.constants.ts
@@ -1,0 +1,12 @@
+// See= https=//developer.mozilla.org/en-US/docs/Web/HTTP/Methods
+export enum HTTP_METHODS {
+    GET = 'GET',
+    POST = 'POST',
+    PUT = 'PUT',
+    DELETE = 'DELETE',
+    PATCH = 'PATCH',
+    OPTIONS = 'OPTIONS',
+    HEAD = 'HEAD',
+    CONNECT = 'CONNECT',
+    TRACE = 'TRACE'
+}