+++ monkey/monkey_island/cc/next_ui/src/constants/paths.constants.tsx
@@ -1,0 +1,10 @@
+export enum PATHS {
+    HOME = '/home',
+    MAP = '/map',
+    EVENTS = '/events',
+    SIGN_IN = '/signin',
+    SIGN_UP = '/signup',
+    ROOT = '/'
+}
+
+export const AUTHENTICATION_PATHS = [PATHS.SIGN_IN, PATHS.SIGN_UP];