+++ monkey/monkey_island/cc/ui/src/components/pages/RunMonkeyPage/commands/local_linux_wget.js
@@ -1,10 +1,10 @@
 import { AGENT_OTP_ENVIRONMENT_VARIABLE } from './consts';
 
-export default function generateLocalLinuxWget(ip, username, otp) {
-  let command = `wget --no-check-certificate https://${ip}:5000/api/agent-binaries/`
+export default function generateLocalLinuxWget(ip, port, username, otp) {
+  let command = `wget --no-check-certificate https://${ip}:${port}/api/agent-binaries/`
     + `linux -O ./monkey-linux-64; `
     + `chmod +x monkey-linux-64; `
-    + `${AGENT_OTP_ENVIRONMENT_VARIABLE}=${otp} ./monkey-linux-64 m0nk3y -s ${ip}:5000`;
+    + `${AGENT_OTP_ENVIRONMENT_VARIABLE}=${otp} ./monkey-linux-64 m0nk3y -s ${ip}:${port}`;
 
   if (username != '') {
     command = `su - ${username} -c "${command}"`;